Typescriptify the util directory
