--- v3/mathjax3-es6/util/AsyncLoad.d.ts
@@ -1,1 +1,0 @@
-export function AsyncLoad(name: string): Promise<any>;
