--- v3/mathjax3-es6/util/AsyncLoad.js
@@ -1,3 +1,0 @@
-export function AsyncLoad(name) {
-  return System.import(name,__moduleName);
-}
