--- v3/mathjax3-es6/util/FunctionList.d.ts
@@ -1,6 +1,0 @@
-import {PrioritizedList} from "./PrioritizedList.js";
-
-export class FunctionList extends PrioritizedList {
-    Execute(...data: any): boolean;
-    asyncExecute(...data: any): void;
-};
