--- v3/mathjax3-es6/util/PrioritizedList.d.ts
@@ -1,6 +1,0 @@
-export class PrioritizedList {
-    items: any[];
-    Add(item: any, priority: number): any;
-    Remove(item: any): void;
-}
-
