--- v3/mathjax3-es6/util/Retries.d.ts
@@ -1,2 +1,0 @@
-export function HandleRetriesFor(code: Function): void;
-export function RetryAfter(promise: Promise<any>): void;
