+++ v3/mathjax3-ts/util/AsynchLoad.ts
@@ -1,0 +1,6 @@
+declare var System: {import: Function};
+declare var __moduleName: string;
+
+export function AsyncLoad(name: string) {
+    return System.import(name,__moduleName);
+}
